+++ Makefile
@@ -138,7 +138,7 @@ unexport INCLUDES
 INCLUDES=-I.
 
 
-OBJS=board.o board_undo.o gtp.o move.o ownermap.o pattern3.o pattern.o patternsp.o patternprob.o playout.o probdist.o random.o stone.o timeinfo.o network.o fbook.o chat.o
+OBJS=board.o gtp.o move.o ownermap.o pattern3.o pattern.o patternsp.o patternprob.o playout.o probdist.o random.o stone.o timeinfo.o network.o fbook.o chat.o
 ifdef DCNN
 	OBJS+=dcnn.o
 endif
